refactor: ui 로직 분류하도록 리팩토링
