+++ src/main/kotlin/blackjack/view/InputView.kt
@@ -1,8 +1,8 @@
 package blackjack.view
 
 import blackjack.model.Player
-import blackjack.model.user.UserDecision
-import blackjack.model.user.UserInputValidator
+import blackjack.view.user.UserDecision
+import blackjack.view.user.UserInputValidator
 
 object InputView {
     private const val INPUT_MESSAGE_PLAYER_NAMES = "게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)"
